Change the keystroke packet

+++ common/src/KeyStroke.cpp
@@ -1,13 +1,13 @@
 #include "KeyStroke.hpp"
 
 KeyStroke::KeyStroke()
-	: APacket(APacket::PacketType::KEYSTROKES), _now(static_cast<unsigned long>(std::time(nullptr))), _data()
+	: APacket(APacket::PacketType::KEYSTROKES), _now(std::time(nullptr)), _data()
 {
 
 }
 
 KeyStroke::KeyStroke(const std::string &data)
-	: APacket(APacket::PacketType::KEYSTROKES), _now(static_cast<unsigned long>(std::time(nullptr))), _data(data)
+	: APacket(APacket::PacketType::KEYSTROKES), _now(std::time(nullptr)), _data(data)
 {
 
 }
@@ -31,14 +31,14 @@ std::vector<uint8_t> KeyStroke::to_bytes_body() const
 void KeyStroke::to_readable_body(IReadable &parser) const
 {
 	parser.put("Time", _now);
-	parser.put("Size", static_cast<int>(_data.size()));
+	parser.put("Size", static_cast<uint16_t>(_data.size()));
 	parser.put("Data", _data);
 }
 
 void KeyStroke::from_bytes_body(const std::vector<uint8_t> &bytes)
 {
     std::size_t pos = 1;
-    unsigned int size = 0;
+	short size = 0;
 
 	get_bytes(bytes, pos, _now);
     get_bytes(bytes, pos, size);
@@ -56,6 +56,6 @@ void KeyStroke::from_readable_body(IReadable &parser)
 	parser.get("Size", size);
 	parser.get("Data", _data);
     
-    if (static_cast<unsigned int>(size) != _data.size())
+    if (static_cast<short>(size) != _data.size())
         throw std::invalid_argument("Error while parsing packet");
 }
